@@ -2,7 +2,7 @@
 /// <reference path="../gulp/gulp.d.ts" />
 
 import gulp = require('gulp');
-import merge2 = require('merge2');
+import * as merge2 from "merge2"; 
 
 gulp.task('app-js', () =>
   merge2(
